Moved kernel hacks into their own file.
