+++ driver/ppm_fillers.h
@@ -10,8 +10,6 @@ or GPL2.txt for full copies of the license.
 #ifndef PPM_FILLERS_H_
 #define PPM_FILLERS_H_
 
-#include <linux/version.h>
-
 /* This is described in syscall(2). Some syscalls take 64-bit arguments. On
  * arches that have 64-bit registers, these arguments are shipped in a register.
  * On 32-bit arches, however, these are split between two consecutive registers,
@@ -28,25 +26,6 @@ or GPL2.txt for full copies of the license.
 #define _64BIT_ARGS_SINGLE_REGISTER
 #endif /* __x86_64__ */
 #endif /* __KERNEL__ */
-
-// probe_kernel_read() only added in kernel 2.6.26, name changed in 5.8.0
-#if LINUX_VERSION_CODE < KERNEL_VERSION(2, 6, 26)
-static inline long copy_from_kernel_nofault(void *dst, const void *src, size_t size)
-{
-	long ret;
-	mm_segment_t old_fs = get_fs();
-
-	set_fs(KERNEL_DS);
-	pagefault_disable();
-	ret = __copy_from_user_inatomic(dst, (__force const void __user *)src, size);
-	pagefault_enable();
-	set_fs(old_fs);
-
-	return ret ? -EFAULT : 0;
-}
-#elif LINUX_VERSION_CODE < KERNEL_VERSION(5, 8, 0)
-#define copy_from_kernel_nofault probe_kernel_read
-#endif
 
 #define FILLER_LIST_MAPPER(FN)			\
 	FN(sys_autofill)			\
